Change historical batch sizes
Changed so that *half* the fake examples seen are historical
examples. Previously, the number of fake examples depended on
the *other* batch size (e.g. fake x batch size depended on y
batch size) but I think it's more correct for the number
of real and fake x examples seen in a mini batch to be equal.